+++ client/src/app/components/resumen/resumen-general.controller.js
@@ -26,8 +26,10 @@
             total_capacitandos   : 0,
             cursos_activados     : 0,
             capacitandos_activos : 0
-        }
+        };
         
+        vm.data1 = [];
+        vm.data2 = [];
 
         inicia();
 
@@ -56,6 +58,23 @@
 
                     vm.calculaDatos();
                 });
+
+                
+                CursosOficiales.cursos_mes({
+                  anio: vm.anio,
+                  meses: '7,8',
+                  filter: {
+                        order: ['mes ASC']
+                  }
+                })
+                .$promise
+                .then(function(resp) {
+                    angular.forEach(resp, function(registro) {
+                        vm.data1.push([gd(registro.anio, registro.mes, 1), registro.num_cursos]);
+                        vm.data2.push([gd(registro.anio, registro.mes, 1), registro.num_personas]);
+                    });
+                });
+
         }
 
 
@@ -63,6 +82,7 @@
             
             var condicion_unidad;
             var condicion_trimestre;
+            var meses = [];
 
             if(vm.unidadSeleccionada.idUnidadAdmtva == 0)
                 condicion_unidad = {idUnidadAdmtva : {gt:1}};
@@ -80,24 +100,28 @@
                 var fecha_inicio = new Date(parseInt(vm.anio), 0, 1, 0, 0, 0, 0);
                 var fecha_fin    = new Date(parseInt(vm.anio), 2, 31, 23, 59, 59, 999);
                 condicion_trimestre = {fechaInicio: {between: [fecha_inicio,fecha_fin]} };
+                meses = [1,2,3];
             }
             else if(parseInt(vm.trimestre) == 2)
             {
                 var fecha_inicio = new Date(parseInt(vm.anio), 3, 1, 0, 0, 0, 0);
                 var fecha_fin    = new Date(parseInt(vm.anio), 5, 30, 23, 59, 59, 999);
                 condicion_trimestre = {fechaInicio: {between: [fecha_inicio,fecha_fin]} };
+                meses = [4,5,6];
             }
             else if(parseInt(vm.trimestre) == 3)
             {
                 var fecha_inicio = new Date(parseInt(vm.anio), 6, 1, 0, 0, 0, 0);
                 var fecha_fin    = new Date(parseInt(vm.anio), 8, 30, 23, 59, 59, 999);
                 condicion_trimestre = {fechaInicio: {between: [fecha_inicio,fecha_fin]} };
+                meses = [7,8,9];
             }
             else if(parseInt(vm.trimestre) == 4)
             {
                 var fecha_inicio = new Date(parseInt(vm.anio), 9, 1, 0, 0, 0, 0);
                 var fecha_fin    = new Date(parseInt(vm.anio), 11, 31, 23, 59, 59, 999);
                 condicion_trimestre = {fechaInicio: {between: [fecha_inicio,fecha_fin]} };
+                meses = [10,11,12];
             }
             else if(parseInt(vm.trimestre) >= 11)
             {
@@ -107,6 +131,7 @@
 
                 var fecha_fin    = new Date(parseInt(vm.anio), (parseInt(vm.trimestre)-11), numDias, 23, 59, 59, 999);
                 condicion_trimestre = {fechaInicio: {between: [fecha_inicio,fecha_fin]} };
+                meses = [(fecha_inicio.getMonth()+1)];
             }
 
             Capacitandos.count({
@@ -180,8 +205,146 @@
                 
             });
 
+            
+            vm.data3 = [];
 
+            if(parseInt(vm.trimestre) > 0)
+            {
+                    CursosOficiales.cursos_mes({
+                      anio: vm.anio,
+                      meses: meses.toString(),
+                      filter: {
+                            order: ['mes ASC']
+                      }
+                    })
+                    .$promise
+                    .then(function(resp) {
+                        angular.forEach(resp, function(registro) {
+                            vm.data3.push([gd(registro.anio, registro.mes, 1), registro.num_personas]);
+                        });
+                    });
+
+            }
+
+
+            var dataset = [
+                {
+                    label: "Personas inscritas",
+                    grow:{stepMode:"linear"},
+                    data: vm.data2,
+                    color: "#1ab394",
+                    bars: {
+                        show: true,
+                        align: "center",
+                        barWidth: 30 * 24 * 60 * 60 * 600,
+                        lineWidth: 0
+                    }
+
+                },
+                {
+                    label: "Meses seleccionados",
+                    grow:{stepMode:"linear"},
+                    data: vm.data3,
+                    color: "#FF0000",
+                    bars: {
+                        show: true,
+                        align: "center",
+                        barWidth: 30 * 24 * 60 * 60 * 600,
+                        lineWidth: 0
+                    }
+
+                },
+                {
+                    label: "Cursos impartidos",
+                    grow:{stepMode:"linear"},
+                    data: vm.data1,
+                    yaxis: 2,
+                    color: "#1C84C6",
+                    lines: {
+                        lineWidth: 1,
+                        show: true,
+                        fill: true,
+                        fillColor: {
+                            colors: [
+                                {
+                                    opacity: 0.2
+                                },
+                                {
+                                    opacity: 0.2
+                                }
+                            ]
+                        }
+                    }
+                }
+            ];
+
+
+            var options = {
+                grid: {
+                    hoverable: true,
+                    clickable: true,
+                    tickColor: "#d5d5d5",
+                    borderWidth: 0,
+                    color: '#d5d5d5'
+                },
+                colors: ["#1ab394", "#464f88"],
+                tooltip: true,
+                xaxis: {
+                    mode: "time",
+                    timeformat: "%b",
+                    tickSize: [1, "month"],
+                    monthNames: ["Ene", "Feb", "Mar", "Abr", "May", "Jun", "Jul", "Ago", "Sep", "Oct", "Nov", "Dic"],
+                    tickLength: 0,
+                    axisLabel: "Mes",
+                    axisLabelUseCanvas: true,
+                    axisLabelFontSizePixels: 12,
+                    axisLabelFontFamily: 'Arial',
+                    axisLabelPadding: 10,
+                    color: "#d5d5d5"
+                },
+                yaxes: [
+                    {
+                        position: "left",
+                        //max: 25,
+                        tickDecimals: 0,
+                        color: "#d5d5d5",
+                        axisLabelUseCanvas: true,
+                        axisLabelFontSizePixels: 12,
+                        axisLabelFontFamily: 'Arial',
+                        axisLabelPadding: 3
+                    },
+                    {
+                        position: "right",
+                        tickDecimals: 0,
+                        color: "#d5d5d5",
+                        axisLabelUseCanvas: true,
+                        axisLabelFontSizePixels: 12,
+                        axisLabelFontFamily: ' Arial',
+                        axisLabelPadding: 67
+                    }
+                ],
+                legend: {
+                    noColumns: 1,
+                    labelBoxBorderColor: "#d5d5d5",
+                    position: "ne",
+                    margin: [-170,15]
+                }
+
+            };
+
+            /**
+             * Definition of variables
+             * Flot chart
+             */
+            vm.flotData = dataset;
+            vm.flotOptions = options;
         }
+
+
+        function gd(year, month, day) {
+            return new Date(year, month - 1, day).getTime();
+        }
+
 
     };