+++ client/src/assets/libs/angular-flot/CHANGELOG.md
@@ -3,6 +3,27 @@
 All notable changes to this project are documented in this file.
 
 
+## 0.0.15 - 2015-11-24
+
+* Use `$timeout` instead of `scope.$apply` in event handlers. See also issue #30.
+
+
+## 0.0.13 - 2015-09-17
+
+* Properly destroy the plot when the directive is being destroyed. Contributed by @Botffy.
+
+
+## 0.0.11 - 2015-09-12
+
+* Watch width and height attributes. Contributed by @gudubeth.
+
+
+## 0.0.10 - 2015-07-23
+
+* Prevent double initialization when both a dataset and plot options are given. Thanks to @liuce for
+  the patch.
+
+
 ## 0.0.9 - 2015-06-30
 
 * Switch to semistandard-format (that performs semicolon insertion) to reformat source code since